Use explicit xarray's indexes: Delete UgridDataArray, UgridDataset
